Cultural storyteller app full featured/requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
plotly>=5.17.0
Pillow>=10.0.0
opencv-python>=4.8.0
pyttsx3>=2.90
edge-tts>=6.1.9
openai>=1.3.0
sqlite3
hashlib2>=1.3.1
streamlit-webrtc>=0.47.1
av>=10.0.0
numpy>=1.24.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-js-eval>=0.1.5
streamlit-lottie>=0.0.5
asyncio
aiortc>=1.5.0
streamlit-modal>=0.1.0
streamlit-aggrid>=0.3.4
streamlit-elements>=0.1.0